renamed training_algorithms to execution_algorithms
