+++ include/lbann/execution_algorithms/training_algorithm.hpp
@@ -37,7 +37,7 @@ namespace lbann {
 // Forward-declare this.
 class execution_context;
 
-/** Base class for LBANN training_algorithms. */
+/** Base class for LBANN execution_algorithms. */
 class training_algorithm {
 public:
