+++ src/CMakeLists.txt
@@ -19,7 +19,7 @@ add_subdirectory(objective_functions)
 add_subdirectory(optimizers)
 add_subdirectory(proto)
 add_subdirectory(trainers)
-add_subdirectory(training_algorithms)
+add_subdirectory(execution_algorithms)
 add_subdirectory(transforms)
 add_subdirectory(utils)
 add_subdirectory(weights)
